+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-20T17:38:29.061Z",
+    "2026-05-21T05:31:25.098Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {