+++ packages/orq-rc/RELEASES.md
@@ -9369,3 +9369,13 @@ Based on:
 - [typescript v4.10.0-rc.23] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.23 - packages/orq-rc
+
+## 2026-05-21 05:34:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.4 (2.884.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.24] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.24 - packages/orq-rc